Frontend validation

+++ routes/upload.js
@@ -5,6 +5,7 @@ const fs = require('fs');
 const User = require('../models/user');
 const path = require('path');
 const validator = require('validator');
+const request = require('request');
 
 router.post('/', isLoggedIn, (req, res) => {
 
@@ -13,43 +14,55 @@ router.post('/', isLoggedIn, (req, res) => {
 
 	form.parse(req, (err, fields, files) => {
 
-		updateUser(req.user, fields, (err, user) => {
-			if(err) throw err;
+		let captcha = fields['g-recaptcha-response'];
+		if(captcha === undefined || captcha === '' || captcha === null)
+			 return res.redirect('/');
 
-			console.log(req.user.name + ': ' + files['filetoupload']['type']);
-			if(types.indexOf(files['filetoupload']['type']) === -1) res.sendfile('public/success-details.html');
-			else if(files['filetoupload']['size'] > 0) {	
-				let oldPath = files.filetoupload.path;
-				let dir = __dirname + '/../uploads/' + req.user.nameConst + ' - ' + req.user._id;
-				let extention = path.extname(files.filetoupload.name);
-				let newPath = dir + '/' + req.user.nameConst + extention;
+		const secretKey = '6LdALDYUAAAAAKwp_nPEq9MsWMGzLMxyNMTUwMCm';
 
-				if(!fs.existsSync(dir)) {
-					fs.mkdir(dir, err => {
-						if(err) throw err;
-						uploadFile(oldPath, newPath, err => {
+		const verifyUrl = `https://google.com/recaptcha/api/siteverify?secret=${secretKey}&response=${captcha}&remoteip=${req.connection.remoteAddress}`;
+
+		request(verifyUrl, (err, response, body) => {
+			body = JSON.parse(body);
+			if(body.success !== undefined && !body.success)
+				return res.redirect('/');
+	
+			updateUser(req.user, fields, (err, user) => {
+				if(err) throw err;
+				if(types.indexOf(files['filetoupload']['type']) === -1) res.sendfile('public/success-details.html');
+				else if(files['filetoupload']['size'] > 0) {	
+					let oldPath = files.filetoupload.path;
+					let dir = __dirname + '/../uploads/' + req.user.nameConst + ' - ' + req.user._id;
+					let extention = path.extname(files.filetoupload.name);
+					let newPath = dir + '/' + req.user.nameConst + extention;
+
+					if(!fs.existsSync(dir)) {
+						fs.mkdir(dir, err => {
 							if(err) throw err;
-							User.findOneAndUpdate({_id: req.user._id}, {submitted: true}, (err, user) =>{
-							 	if(err) throw err;
-								res.sendfile('public/success-file.html');
-							});
-						});
-					});
-				} else {
-					fs.readdir(dir, (err, files) => {
-						if (err) throw err;
-						let file = files[0];
-						fs.unlink(path.join(dir, file), err => {
-							if (err) throw err;
 							uploadFile(oldPath, newPath, err => {
 								if(err) throw err;
-								else res.sendfile('public/success-file.html');
+								User.findOneAndUpdate({_id: req.user._id}, {submitted: true}, (err, user) =>{
+									if(err) throw err;
+									res.sendfile('public/success-file.html');
+								});
 							});
 						});
-					});
-				}
-			} else res.sendfile('public/success-details.html');
+					} else {
+						fs.readdir(dir, (err, files) => {
+							if (err) throw err;
+							let file = files[0];
+							fs.unlink(path.join(dir, file), err => {
+								if (err) throw err;
+								uploadFile(oldPath, newPath, err => {
+									if(err) throw err;
+									else res.sendfile('public/success-file.html');
+								});
+							});
+						});
+					}
+				} else res.sendfile('public/success-details.html');
 
+			});
 		});
 	});
